+++ src/main/java/com/buglife/sdk/BugReporterImpl.java
@@ -30,9 +30,7 @@ import com.buglife.sdk.reporting.SubmitReportService;
 
 import org.json.JSONException;
 
-import java.io.File;
-
-class BugReporterImpl implements BugReporter {
+final class BugReporterImpl implements BugReporter {
     private final Context mContext;
 
     BugReporterImpl(Context context) {
@@ -55,9 +53,7 @@ class BugReporterImpl implements BugReporter {
     private void reportWithLegacy(Report report) {
         try {
             String jsonReport = report.toJSON().toString();
-            File file = FileUtils.getReportsCacheFile(mContext);
-            FileUtils.appendLineToFile(jsonReport, file);
-            SubmitReportLegacyService.start(mContext);
+            SubmitReportLegacyService.start(mContext, jsonReport);
         } catch (JSONException e) {
             Log.e("Failed to encode report!", e);
         }
