--- src/main/java/com/buglife/sdk/FileUtils.java
@@ -1,102 +1,0 @@
-/*
- * Copyright (C) 2017 Buglife, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.buglife.sdk;
-
-import android.content.Context;
-import android.support.annotation.Nullable;
-
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.Closeable;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-
-public class FileUtils {
-    private FileUtils() {/* No instances */}
-
-    public static File getReportsCacheFile(Context context) {
-        File dir = context.getFilesDir();
-        return new File(dir, "reports");
-    }
-
-    public static void appendLineToFile(String string, File file) {
-        BufferedWriter writer = null;
-        try {
-            writer = new BufferedWriter(new FileWriter(file, true));
-            writer.write(string);
-            writer.newLine();
-        } catch (FileNotFoundException error) {
-            error.printStackTrace();
-        } catch (IOException error) {
-            error.printStackTrace();
-        } finally {
-            closeQuietly(writer);
-        }
-    }
-
-    public static void writeLinesToFile(List<String> lines, File file) {
-        BufferedWriter writer = null;
-        try {
-            writer = new BufferedWriter(new FileWriter(file));
-            for (String line : lines) {
-                writer.write(line);
-                writer.newLine();
-            }
-        } catch (FileNotFoundException error) {
-            error.printStackTrace();
-        } catch (IOException error) {
-            error.printStackTrace();
-        } finally {
-            closeQuietly(writer);
-        }
-    }
-
-    public static List<String> readLinesFromFile(File file) {
-        List<String> output = new ArrayList<>();
-        BufferedReader reader = null;
-        try {
-            reader = new BufferedReader(new FileReader(file));
-            String line;
-            while ((line = reader.readLine()) != null) {
-                output.add(line);
-            }
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            closeQuietly(reader);
-        }
-        return output;
-    }
-
-    private static void closeQuietly(@Nullable Closeable closeable) {
-        if (closeable != null) {
-            try {
-                closeable.close();
-            } catch (IOException error) {
-                // Ignore
-            }
-        }
-    }
-}
